sdk/python/requirements.txt
requests>=2.25.0
pynacl>=1.5.0
blake3>=0.3.0
pytest>=7.0
pytest-cov>=4.0
